+++ src/Engine/SoundEngine.cpp
@@ -23,7 +23,6 @@
 #ifdef MCENGINE_FEATURE_MULTITHREADING
 
 #include <mutex>
-#include "WinMinGW.Mutex.h"
 
 #endif
 
@@ -262,7 +261,7 @@ SoundEngine::SoundEngine()
         engine->shutdown();
         return;
     }
-    if (!(m_BASS_FX_TempoCreate = (unsigned int (*)(unsigned int, unsigned int))dlsym(bassfx_handle, "BASS_FX_TempoCreate")))
+    if (!(m_BASS_FX_TempoCreate = (long unsigned int (*)(long unsigned int, long unsigned int))dlsym(bassfx_handle, "BASS_FX_TempoCreate")))
     {
         engine->showMessageErrorFatal("failed to find BASS_FX_TempoCreate in libbass_fx.so: %s\n", dlerror());
         engine->shutdown();
